+++ packages/demo/view-libs/react.next@19.x/app/todo-app/components/TodoAdder/index.tsx
@@ -1,11 +1,11 @@
 export const TodoAdder = reaxper(() => {
-	const { AddNew_Store , setAddInput , addNew } = reaxel_AddNew();
+	const { setAddInput , addNew } = reaxel_AddNew();
 	
 	return <div className={style['todo-adder']}>
 		<input
 			type = "text"
 			placeholder = "type something here.."
-			value = { AddNew_Store.addInputValue }
+			value = { reaxel_AddNew.store.addInputValue }
 			onInput = { ( event: React.ChangeEvent<HTMLInputElement> ) => {
 				setAddInput(event.target.value);
 			} }
@@ -21,12 +21,12 @@ export const TodoAdder = reaxper(() => {
 				addNew();
 				setAddInput('');
 			} }
-			disabled = { !AddNew_Store.addInputValue }
+			disabled = { !reaxel_AddNew.store.addInputValue }
 		>{ "add to ToDo" }</button>
 	</div>;
 });
 
 import React from 'react';
 import { reaxper } from 'reaxes-react';
-import { reaxel_AddNew } from 'reaxels/todo-app/add-new';
+import { reaxel_AddNew } from '@/reaxels/todo-app/add-new';
 import style from './style.module.css';