+++ packages/demo/view-libs/react.next@19.x/app/todo-app/components/TodoList/index.tsx
@@ -1,9 +1,9 @@
 export const TodoList = reaxper(() => {
-	const { ListFilter_Store , setFilterCriteria } = reaxel_ListFilter();
+	const { setFilterCriteria } = reaxel_ListFilter();
 	
 	return <div className={style['todo-list']}>
 		<ul>
-			{ ListFilter_Store.filteredTodoList.map(( item,index,arr ) => <React.Fragment key = { item.id }>
+			{ reaxel_ListFilter.store.filteredTodoList.map(( item,index,arr ) => <React.Fragment key = { item.id }>
 				<TodoListItem
 					todoItem = { item }
 				/>
@@ -17,5 +17,5 @@ import React from 'react';
 import { reaxper } from 'reaxes-react';
 import { ListFilter } from '../ListFilter';
 import { TodoListItem } from '../TodoListItem';
-import { reaxel_ListFilter } from 'reaxels/todo-app/filter-list';
+import { reaxel_ListFilter } from '@/reaxels/todo-app/filter-list';
 import style from './style.module.css';