+++ packages/demo/view-libs/react.next@19.x/libs/reaxes-react/types/enhancer/withHoC.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="react" />
-export declare function withHoC<T extends (React.Component & React.FC)>(OriginalComponent: T): React.FunctionComponent<{}>;