+++ packages/demo/view-libs/react.next@19.x/libs/reaxes-react/types/index.d.ts
@@ -1,3 +1,0 @@
-export * from './reaxper';
-export * from './Reaxlass';
-export * from 'reaxes';