+++ packages/demo/view-libs/react.next@19.x/libs/reaxes-utils/types/crayon.utility.d.ts
@@ -1,9 +1,0 @@
-type crayon = ((css: Partial<CSSStyleDeclaration>) => (...logs: any[]) => void) & {
-    [key in "warn" | "info" | "log" | "error" | "debug" | "trace"]: ((...any: any[]) => any) & {
-        [key in string]: (...any: any[]) => any;
-    };
-} & {
-    [key in string]: (...any: any[]) => any;
-};
-export declare const crayon: crayon;
-export {};