+++ packages/demo/view-libs/react.next@19.x/libs/reaxes-utils/types/getNestedValue.utility.d.ts
@@ -1,2 +1,0 @@
-declare function getNestedValue(targetObj: object, keys: any[]): any;
-export default getNestedValue;