+++ packages/demo/view-libs/react.next@19.x/libs/reaxes-utils/types/throttle.utility.d.ts
@@ -1,6 +1,0 @@
-interface IOptions {
-    leading?: boolean;
-    trailing?: boolean;
-}
-export default function throttle(func: any, wait: any, options: IOptions): () => void;
-export {};