+++ packages/reaxes-utils/src/xPromise.ts
@@ -3,24 +3,27 @@
  * @description 生成一个pending状态的promise
  */
 
-export const xPromise = <T = any>(
-	callback?: (resolve: Function, reject: Function) => any,
+export const xPromise = <T>(
+	callback?: (
+		resolve: (value: T | PromiseLike<T>) => void,
+		reject: (reason?: any) => void) => void
 ): Promise<T> & {
-	resolve: (value?: any) => void;
+	resolve: (value: T | PromiseLike<T>) => void;
 	reject: (reason?: any) => void;
 } => {
-	let resolve: (value?: any) => void, reject: (reason?: any) => void;
+	let resolve!: (value: T | PromiseLike<T>) => void;
+	let reject!: (reason?: any) => void;
+	
 	const promise = new Promise<T>(($resolve, $reject) => {
 		resolve = $resolve;
 		reject = $reject;
 		typeof callback === 'function' && callback($resolve, $reject);
 	});
-	Object.assign(promise, {
+	
+	return Object.assign(promise, {
 		resolve,
 		reject,
 	});
-	
-	return promise as Promise<T> & {resolve; reject};
 };
 
 export type XPromise<T = any> = Promise<T> & {resolve:(value:T) => void; reject:(reason:any) => void}